Revert "fix: testing new logic"
This reverts commit 6d18fd28afe37e7879bf127db25065a157a85a67.

+++ persistence/src/main/java/com/bueno/persistence/repositories/RankBotsRepositoryImpl.java
@@ -72,9 +72,8 @@ public class RankBotsRepositoryImpl implements RankBotsRepository {
         try {
             if (botRankInfoDtos.isEmpty()) throw new Exception("BotRankInfoDto is empty");
 
-//            if (botRankInfoDtos.size() < findAll().size()) {
-//                throw new Exception("BotRankInfoDto size is smaller than before, this is just a security feature");
-//            }
+            if (botRankInfoDtos.size() < findAll().size())
+                throw new Exception("BotRankInfoDto size is smaller than before, this is just a security feature");
 
             deleteAll();
             saveAll(botRankInfoDtos);
